# Final/swagger.json
{
    "swagger": "2.0",
    "basePath": "/",
    "paths": {
...
        "/movies/{name}": {
            "parameters": [
                {
                    "in": "path",
                    "description": "The name of movie",
                    "name": "name",
                    "required": true,
                    "type": "string"
                }
            ],
            "get": {
                "responses": {
                    "200": {
                        "description": "Successful"
                    },
                    "404": {
                        "description": "Movie was not found"
                    }
                },
                "description": "Get a Movie by its title",
                "operationId": "get_movie",
                "tags": [
                    "Movies"
                ]
            }
        },
...
    },
    "info": {
        "title": "Movie Dataset",
        "version": "1.0",
        "description": "This is the movie recommendation system"
    },
    "produces": [
        "application/json"
    ],
    "consumes": [
        "application/json"
    ],
    "securityDefinitions": {
        "API-KEY": {
            "type": "apiKey",
            "in": "header",
            "name": "AUTH-TOKEN"
        }
    },
    "security": [
        {
            "API-KEY": []
        }
    ],
...
}